Fixed main.py

+++ src/main.py
@@ -63,7 +63,7 @@ def multiplication():
 	n1 = text1.get()
 	n2 = text2.get()
 
-	r = int(n1) * int(n2)
+	r = float(n1) * float(n2)
 	text3.delete(0,'end')
 	text3.insert(0,r)
 	label = tk.Label(screen,text=f'You have multiplied {n1} by {2}')
